+++ packages/sdk/src/types/attributes/requests/GetSharedVersionsOfRepositoryAttributeRequest.ts
@@ -1,0 +1,4 @@
+export interface GetSharedVersionsOfRepositoryAttributeRequest {
+    peers?: string[];
+    onlyLatestVersions?: boolean;
+}