+++ test/messages.test.ts
@@ -1,6 +1,7 @@
 import { ConnectorClient } from "@nmshd/connector-sdk";
 import { Launcher } from "./lib/Launcher";
 import { QueryParamConditions } from "./lib/QueryParamConditions";
+import { getTimeout } from "./lib/setTimeout";
 import { establishRelationship, exchangeMessage, getRelationship, syncUntilHasMessages, uploadFile } from "./lib/testUtils";
 import { ValidationSchema } from "./lib/validation";
 
@@ -15,7 +16,7 @@ beforeAll(async () => {
 
     const relationship = await getRelationship(client1);
     connector2Address = relationship.peer;
-}, 30000);
+}, getTimeout(30000));
 
 afterAll(() => launcher.stop());