+++ test/outgoingRequests.test.ts
@@ -2,13 +2,14 @@ import { ConnectorClient, ConnectorRequestStatus } from "@nmshd/connector-sdk";
 import { DateTime } from "luxon";
 import { Launcher } from "./lib/Launcher";
 import { QueryParamConditions } from "./lib/QueryParamConditions";
+import { getTimeout } from "./lib/setTimeout";
 import { ValidationSchema } from "./lib/validation";
 
 const launcher = new Launcher();
 let client1: ConnectorClient;
 let client2: ConnectorClient;
 
-beforeAll(async () => ([client1, client2] = await launcher.launch(2)), 30000);
+beforeAll(async () => ([client1, client2] = await launcher.launch(2)), getTimeout(30000));
 afterAll(() => launcher.stop());
 
 describe("Outgoing Requests", () => {
@@ -16,7 +17,7 @@ describe("Outgoing Requests", () => {
         const response = await client2.outgoingRequests.createRequest({
             content: {
                 items: [{ "@type": "ReadAttributeRequestItem", mustBeAccepted: false, query: { "@type": "IdentityAttributeQuery", valueType: "Surname" } }],
-                expiresAt: DateTime.now().plus({ hour: 1 }).toISO()!
+                expiresAt: DateTime.now().plus({ hour: 1 }).toISO()
             },
             peer: (await client1.account.getIdentityInfo()).result.address
         });
@@ -34,7 +35,7 @@ describe("Outgoing Requests", () => {
         const response = await client1.outgoingRequests.createRequest({
             content: {
                 items: [{ "@type": "ReadAttributeRequestItem", mustBeAccepted: false, query: { "@type": "IdentityAttributeQuery", valueType: "Surname" } }],
-                expiresAt: DateTime.now().plus({ hour: 1 }).toISO()!
+                expiresAt: DateTime.now().plus({ hour: 1 }).toISO()
             },
             peer: (await client2.account.getIdentityInfo()).result.address
         });